requirements.txt
streamlit
fastapi
uvicorn
networkx
numpy
sentence-transformers
torch
torch-geometric
datasets
langchain-core
langchain-text-splitters
openai
python-dotenv
tqdm
